Bug fix:  Remove halo.
The edge was not perfect.
It was most obvious when the original image was white
and the background that we are displaying against was white.
The edge should disappear because it's white on white.
The details vary, but there was often a halo around the edge.

I'm using a new method of doing the same thing.
So now the edge of the image looks perfect.